Refactor backup code. (#4825)
- Keep command line code in ee/backup.
- Move backup code to the worker package. This is needed so that a
  restore can assign uids. Currently that's not possible due to cyclic
  depenedencies.
- Move backup tests to systest/backup.
