+++ systest/backup/filesystem/backup_test.go
@@ -34,8 +34,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 
-	"github.com/dgraph-io/dgraph/ee/backup"
 	"github.com/dgraph-io/dgraph/testutil"
+	"github.com/dgraph-io/dgraph/worker"
 	"github.com/dgraph-io/dgraph/x"
 )
 
@@ -265,7 +265,7 @@ func runRestore(t *testing.T, backupLocation, lastDir string, commitTs uint64) m
 	require.NoError(t, os.RemoveAll(restoreDir))
 
 	t.Logf("--- Restoring from: %q", backupLocation)
-	_, err := backup.RunRestore("./data/restore", backupLocation, lastDir)
+	_, err := worker.RunRestore("./data/restore", backupLocation, lastDir)
 	require.NoError(t, err)
 
 	for i, pdir := range []string{"p1", "p2", "p3"} {
@@ -297,7 +297,7 @@ func runFailingRestore(t *testing.T, backupLocation, lastDir string, commitTs ui
 	// calling restore.
 	require.NoError(t, os.RemoveAll(restoreDir))
 
-	_, err := backup.RunRestore("./data/restore", backupLocation, lastDir)
+	_, err := worker.RunRestore("./data/restore", backupLocation, lastDir)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "expected a BackupNum value of 1")
 }
